pyproject.toml
[project]
name = "onnx-predict-yolov8"
version = "1.0.3"
description = ""
readme = "README.md"
authors = [
    {name = "Kasper Fromm Pedersen", email = "kasperf@cs.aau.dk"},
    {name = "Kristian Torp", email = "torp@cs.aau.dk"}
]
maintainers = [
    {name = "Kasper Fromm Pedersen", email = "kasperf@cs.aau.dk"},
]
license = { file = "LICENSE"}
keywords = ["ONNX", "YOLOv8", "onnxruntime", "vision"]
requires-python = ">=3.8,<3.11"
dependencies = [
    "numpy>=1.20",
    "Pillow>=9.1.0",
]

[project.urls]
homepage = "https://www.cs.aau.dk/"
repository = "https://github.com/fromm1990/onnx-predict-yolov8"